feat: Add PersonDto for data transfer and update use cases to utilize DTOs
