+++ src/main/java/com/veterinaria/consentidos/features/person/application/usecase/DeletePersonUseCase.java
@@ -3,6 +3,7 @@ package com.veterinaria.consentidos.features.person.application.usecase;
 import com.veterinaria.consentidos.features.person.domain.repository.PersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Use case for deleting a person.
@@ -23,6 +24,7 @@ public class DeletePersonUseCase {
      * @param id the ID of the person to delete
      * @throws IllegalArgumentException if the person is not found
      */
+    @Transactional
     public void execute(Long id) {
         if (!personRepository.findById(id).isPresent()) {
             throw new IllegalArgumentException("Person with ID " + id + " not found");
