+++ src/main/java/com/veterinaria/consentidos/features/person/presentation/controller/PersonController.java
@@ -2,7 +2,7 @@ package com.veterinaria.consentidos.features.person.presentation.controller;
 
 import com.veterinaria.consentidos.core.PagedResult;
 import com.veterinaria.consentidos.features.person.application.command.CreatePersonCommand;
-import com.veterinaria.consentidos.features.person.application.command.PersonDto;
+import com.veterinaria.consentidos.features.person.application.dto.PersonDto;
 import com.veterinaria.consentidos.features.person.application.command.UpdatePersonCommand;
 import com.veterinaria.consentidos.features.person.application.usecase.CreatePersonUseCase;
 import com.veterinaria.consentidos.features.person.application.usecase.DeletePersonUseCase;
